Fix problem in counting links in retweets
In some versions of a retweet, the retweet_count of the status is
the same as the retweet_count of the retweeted_status, which makes the
algorithm weigh some tweets too heavily. Example: wikileaks had a link to
the talk about the transhumanist paradox with 500+ retweets. 340+ of these
retweets had the same number (500+) of retweets. This is possibly because
of the way Twitter deals with retweets (maybe symlink-like constructions?)

Solution:
If status[retweets] == status[retweeted_status][retweets] than we probably
have this problem. In that case, do not count the retweets of the status.